fix(backend): add mypy type annotations to ML and prediction engine
- Add return type annotations to functions in model_loader, trainer, pipeline
- Fix dict type parameters (dict[str, Any] instead of dict)
- Add type: ignore comments for optional ML library interactions
- Fix dataclass field defaults in ensemble_advanced
- Add proper Any typing for SQLAlchemy Base class
- Fix float/int type mismatches in database stats

Remaining mypy issues in data_enrichment, rag_enrichment, and API routes
will be addressed in follow-up work.

+++ backend/src/prediction_engine/models/poisson.py
@@ -260,7 +260,7 @@ class PoissonModel:
         home_scores = 1 - poisson.pmf(0, expected_home)
         away_scores = 1 - poisson.pmf(0, expected_away)
 
-        return home_scores * away_scores
+        return float(home_scores * away_scores)
 
 
 # Default instance
